Install apptainer with setuid not for GitPod

+++ scripts/ubuntu/install_apptainer.sh
@@ -6,10 +6,13 @@ sudo apt update
 sudo apt install -y software-properties-common
 sudo add-apt-repository -y ppa:apptainer/ppa
 sudo apt update
-sudo apt install -y apptainer
-sudo add-apt-repository -y ppa:apptainer/ppa
-sudo apt update
-sudo apt install -y apptainer-suid
+# perform setuid installation if not in GitPod
+if [ -z "$(env | grep -E "^GITPOD")" ]; then
+  sudo apt install -y apptainer
+  sudo add-apt-repository -y ppa:apptainer/ppa
+  sudo apt update
+  sudo apt install -y apptainer-suid
+fi
 
 # The if chunk below is meant to run if it's in a 
 # docker container.
@@ -27,6 +30,8 @@ if [ -f /.dockerenv ] | [ -n "$(env | grep -E "^GITPOD|^CODESPACE")" ]; then
   if [ -n "$(env | grep -E "^GITPOD")" ]; then 
     grep -q "bind path = /workspace" /etc/apptainer/apptainer.conf || \
       sudo sed -i "s|bind path = /etc/hosts|bind path = /etc/hosts\nbind path = /workspace|" /etc/apptainer/apptainer.conf
+    grep -q "bind path = /nix" /etc/apptainer/apptainer.conf || \
+      sudo sed -i "s|bind path = /etc/hosts|bind path = /etc/hosts\nbind path = /nix|" /etc/apptainer/apptainer.conf
   elif [ -n "$(env | grep -E "^CODESPACE")" ]; then 
     grep -q "bind path = /workspaces" /etc/apptainer/apptainer.conf || \
       sudo sed -i "s|bind path = /etc/hosts|bind path = /etc/hosts\nbind path = /workspaces|" /etc/apptainer/apptainer.conf
